+++ Kiwano/ui/Menu.cpp
@@ -41,7 +41,7 @@ namespace kiwano
 		return enabled_;
 	}
 
-	int Menu::GetButtonCount() const
+	size_t Menu::GetButtonCount() const
 	{
 		return buttons_.size();
 	}
