+++ Kiwano/utils/ResLoader.cpp
@@ -68,7 +68,7 @@ namespace kiwano
 		return false;
 	}
 
-	int ResLoader::AddFrames(String const& id, Array<Resource> const& images)
+	size_t ResLoader::AddFrames(String const& id, Array<Resource> const& images)
 	{
 		if (images.empty())
 			return 0;
@@ -100,7 +100,7 @@ namespace kiwano
 		return 0;
 	}
 
-	int ResLoader::AddFrames(String const& id, Array<ImagePtr> const& images)
+	size_t ResLoader::AddFrames(String const& id, Array<ImagePtr> const& images)
 	{
 		if (images.empty())
 			return 0;
@@ -114,7 +114,7 @@ namespace kiwano
 		return 0;
 	}
 
-	int ResLoader::AddFrames(String const & id, Resource const & image, int cols, int rows)
+	size_t ResLoader::AddFrames(String const & id, Resource const & image, int cols, int rows)
 	{
 		if (cols <= 0 || rows <= 0)
 			return 0;
@@ -153,7 +153,7 @@ namespace kiwano
 		return 0;
 	}
 
-	int ResLoader::AddFrames(String const & id, Resource const & image, Array<Rect> const & crop_rects)
+	size_t ResLoader::AddFrames(String const & id, Resource const & image, Array<Rect> const & crop_rects)
 	{
 		ImagePtr raw = new (std::nothrow) Image;
 		if (!raw || !raw->Load(LocateRes(image, search_paths_)))
